requirements.txt
fastapi
uvicorn
pillow
torch
transformers
opencv-python
numpy
scikit-learn
aiofiles
